+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `objectstorage` commands (`delete`, `list`, `show`) for Managed object storage management
+
 ## [3.1.0] - 2023-11-06
 
 ### Added